Boxer-Bugs-13 Fixing signatures for com-mouse-xyzed commands.
In coms-oglmouse and popup, a number of `defboxer-command` invocations
for commands primarily starting with `com-mouse-` can be called from
the editor with zero arguments when invoking the default magic mouse
commands, such as `MOUSE-CLICK-ON-BOTTOM-RIGHT` or `COMMAND-MOUSE-CLICK`.

The lower level lisp functions all take window, x, y, and mouse-bp which
can be calculated from the environment in the event that they aren't
passed to the function. Only a few of these invocations had the arguments
set to &optional. This converts the parameters for the rest of them to
&optional, with defaults calculated from the environment in the event
they are invoked by the use of these magic mouse functions.