the-attic Removing with-sprites-hidden macros
with-sprites-hidden and the tooling around it are no longer necessary since
deprecating the options to never render during eval and to only render during
eval in the case of changed graphics.  All that with-sprites-hidden 't' achieved was
to push screen boxes on to a global list and this is what determined whether
rendering should happen when it was set to :changed.

With opengl screen buffer swapping we just always rerender the entire screen now.