bugs-118 Fixing NSAssertMainEventQueueIsCurrentEventQueue error
In the process of reading the preferences before displaying the window
we ran in to this error doing other thread tasks (reading files on disk)
before the application was initialized.  This was because we wanted the
user preferences on window size to be set as capi dimension hints.

In order to force starting up before we display the main window, we can
use convert-to-screen with nil.