refactor the-attic Removing last major sections of deprecated boxer server client code
Removes the last vestiges of the client properties that were kept on the plist
of boxes, most with the prefix of cross-file-*. This includes hooks for it in the
virtual copy, editor, loading/dumping, and other minor bits of architecture. This
removes a lot of older complex code that hasn't been used since probably the Sun
days.

- Completely commenting out client.lisp an clientmacros.lisp from our asdf modules.
  Will completely archive the files after more testing and usage.
- Moved still necessary items to surf.lisp:
    - defuns no-inferiors-for-file?, storage-chunk-plist-half-length
    - defmethods dump-storage-chunk-plist-items, filename-for-file