refactor Reorganizing the blinkers and boxer canvas
Removing lots of unneeded cruft for the blinker (cursor), and moving it back to the
generic :boxer package. Creating a new canvas class to mix in to the boxer opengl pane
to coalesce global variables that are specific to an individual canvas, starting to
allow for the future possibility of having multiple windows, tabs, divs.