sunrise-67 openGL Framebuffers to back turtle drawing
This is the first major set of work for openGL framebuffers.
There may be other minor cleanups, especially in regards to
resizing graphics and the backing framebuffer textures.

This ticket also removes the old gdispl performance lists,
and adds a new user preference to turn the openGL framebuffers
on and off. Additionally the *features* entry :moderngl has
been removed, and now using openGL 3.2+ is the default.