crash-fix Saving :cached-boxtops with old graphics commands
Working on an issue from the most recent build, where there are old graphics lists that are stored
in box plists under :cached-boxtop that aren't getting translated, so it's still trying to save the old
lower ( >32 ) commands, but are blowing up.

To fix this I put in a check for lower opcodes, and then fetch the higher level graphics-command to
use with dump-gc.  This is ok, after reviewing all the defmethod dump-gc implementations, none of
actually modify the original opcode when saving, and the special ones mostly deal with things like
bitmaps and whatnot. (fonts, colors, bitmaps, alu).