sunrise-112 Fixing stamp-self location
This required a reworking of the graphics display list commands, so stamp-self will work
properly, however, existing draw operations will need to be redrawn for the effects to take
place.

This now always translates the stamp from 0,0 rather than a value which may change depending
on the size of the canvas and other environmental factors. This also adds a  option to
command 37 to pop off the transform matrix that was pushed on the draw stack.